Continued the TUI implementation.
Created the printMap method to print the map as text.
Created the methods to move the armies from the player to the territory.
Created the method to initalize the armies for all the players.
Created the method to let the AI place armies in a random territory.
Created the method to get the free armies of a player.