Added GUI features.
Added the FontManager to keep track of the current font (the custom one,
if available, otherwise the basic Arial).
Added the custom Dialog with the background image because honestly it's
pretty cool.
Added the switch between main menu and setup view.
Added the name dialog in the setup view to ask the player's name.