Did a lot of stuff
Refactored the die folders to use the enumerator name.
Added the WHITE enumerator to the die color.
Changed the roll-your-die screen to choose who started.
Created the right panel to show the players and a small arrow pointing
to the current playing player.
Added a turnlistener to the game in order to notify the GUI when the
turn changes.